컴포넌트 및 템플릿 확장 API 패키지 정리
