+++ sdg/src/main/java/com/shopl/sdg/template/foundation_list/user_list/SDGUserList.kt
@@ -8,9 +8,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.shopl.sdg.component.avatar.SDGAvatarBadge
 import com.shopl.sdg.template.popup.center.SDGCenterPopupButtonOption
+import com.shopl.sdg.template.popup.center.extension.SDGCenterPopup
 import com.shopl.sdg.template.profile.second.SDGSecondProfile
 import com.shopl.sdg.template.profile.second.SDGSecondProfileType
-import com.shopl.sdg.template.util.popup.SDGCenterPopup
 import com.shopl.sdg_common.foundation.SDGColor
 import com.shopl.sdg_common.foundation.spacing.SDGSpacing
 import kotlinx.collections.immutable.PersistentList
