+++ sdg/src/main/java/com/shopl/sdg/template/navi_search_bar/SDGNaviSearchBar.kt
@@ -17,7 +17,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
 import com.shopl.sdg.component.search_bar.capsule.SDGCapsuleSearchType
-import com.shopl.sdg.component.util.search_bar.capsule.SDGCapsuleSearch
+import com.shopl.sdg.component.search_bar.capsule.extension.SDGCapsuleSearch
 import com.shopl.sdg.template.navi_search_bar.preview.SDGSearchNaviPreviewData
 import com.shopl.sdg.template.navi_search_bar.preview.SDGSearchNaviPreviewParameterProvider
 import com.shopl.sdg_common.ext.clickable
