+++ sdg/src/main/java/com/shopl/sdg/template/popup/list/anatomy/SDGListPopupBody.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import com.shopl.sdg.template.util.list_popup_item_ui_state.SDGListPopupItemUiState
+import com.shopl.sdg.template.popup.list.anatomy.extension.SDGListPopupItemUiState
 import com.shopl.sdg_common.ext.clickable
 import com.shopl.sdg_common.foundation.SDGColor
 import com.shopl.sdg_common.foundation.spacing.SDGSpacing.Spacing16
